fix: resolve all TypeScript and ESLint errors
- Replace all 'any' types with proper TypeScript interfaces for Monaco Editor
- Extract useYAMLValidatorSettings hook to separate file to fix react-refresh issues
- Remove unused error variable in yaml-ai-service.ts
- Add proper type definitions for Monaco Editor instances and window.monaco
- Update imports to use the new hook location
- Maintain type safety while eliminating linting errors

All 8 linting errors now resolved:
 TestYAMLValidator.tsx: Fixed Monaco Editor 'any' type
 YAMLEditor.tsx: Fixed Monaco Editor interface types
 YAMLSettingsPanel.tsx: Moved hook to separate file
 useMonacoEditor.ts: Fixed all Monaco Editor 'any' types
 yaml-ai-service.ts: Removed unused error variable

Build and type-check passing successfully.
